Extract the estimated covariance matrices.

Returns a list with K dataframes, each of dimension TxT, corresponding to the covariance matrices estimated for each clone k

Usage

get_covariance_Sigma(x)

Arguments

x

a mvnmm object.

Value

list of the estimated covariance matrices.

Examples

get_covariance_Sigma(x)
#> $C_0
#>              cov_early.L2 cov_early.L1 cov_early.L4 cov_early.L3 cov_mid.L2
#> cov_early.L2      29.3429      0.00000      0.00000      0.00000    0.00000
#> cov_early.L1       0.0000     24.05631      0.00000      0.00000    0.00000
#> cov_early.L4       0.0000      0.00000     16.18587      0.00000    0.00000
#> cov_early.L3       0.0000      0.00000      0.00000     15.53884    0.00000
#> cov_mid.L2         0.0000      0.00000      0.00000      0.00000   22.05584
#> cov_mid.L1         0.0000      0.00000      0.00000      0.00000    0.00000
#> cov_mid.L4         0.0000      0.00000      0.00000      0.00000    0.00000
#> cov_mid.L3         0.0000      0.00000      0.00000      0.00000    0.00000
#> cov_late.L2        0.0000      0.00000      0.00000      0.00000    0.00000
#> cov_late.L1        0.0000      0.00000      0.00000      0.00000    0.00000
#> cov_late.L4        0.0000      0.00000      0.00000      0.00000    0.00000
#> cov_late.L3        0.0000      0.00000      0.00000      0.00000    0.00000
#>              cov_mid.L1 cov_mid.L4 cov_mid.L3 cov_late.L2 cov_late.L1
#> cov_early.L2    0.00000     0.0000    0.00000     0.00000      0.0000
#> cov_early.L1    0.00000     0.0000    0.00000     0.00000      0.0000
#> cov_early.L4    0.00000     0.0000    0.00000     0.00000      0.0000
#> cov_early.L3    0.00000     0.0000    0.00000     0.00000      0.0000
#> cov_mid.L2      0.00000     0.0000    0.00000     0.00000      0.0000
#> cov_mid.L1     22.91318     0.0000    0.00000     0.00000      0.0000
#> cov_mid.L4      0.00000    27.8517    0.00000     0.00000      0.0000
#> cov_mid.L3      0.00000     0.0000   14.62244     0.00000      0.0000
#> cov_late.L2     0.00000     0.0000    0.00000    16.14075      0.0000
#> cov_late.L1     0.00000     0.0000    0.00000     0.00000     67.9152
#> cov_late.L4     0.00000     0.0000    0.00000     0.00000      0.0000
#> cov_late.L3     0.00000     0.0000    0.00000     0.00000      0.0000
#>              cov_late.L4 cov_late.L3
#> cov_early.L2     0.00000     0.00000
#> cov_early.L1     0.00000     0.00000
#> cov_early.L4     0.00000     0.00000
#> cov_early.L3     0.00000     0.00000
#> cov_mid.L2       0.00000     0.00000
#> cov_mid.L1       0.00000     0.00000
#> cov_mid.L4       0.00000     0.00000
#> cov_mid.L3       0.00000     0.00000
#> cov_late.L2      0.00000     0.00000
#> cov_late.L1      0.00000     0.00000
#> cov_late.L4     15.07534     0.00000
#> cov_late.L3      0.00000    49.70778
#> 
#> $C_1
#>              cov_early.L2 cov_early.L1 cov_early.L4 cov_early.L3 cov_mid.L2
#> cov_early.L2     11.06455      0.00000     0.000000      0.00000    0.00000
#> cov_early.L1      0.00000     10.36527     0.000000      0.00000    0.00000
#> cov_early.L4      0.00000      0.00000     8.255685      0.00000    0.00000
#> cov_early.L3      0.00000      0.00000     0.000000      7.48965    0.00000
#> cov_mid.L2        0.00000      0.00000     0.000000      0.00000   10.17361
#> cov_mid.L1        0.00000      0.00000     0.000000      0.00000    0.00000
#> cov_mid.L4        0.00000      0.00000     0.000000      0.00000    0.00000
#> cov_mid.L3        0.00000      0.00000     0.000000      0.00000    0.00000
#> cov_late.L2       0.00000      0.00000     0.000000      0.00000    0.00000
#> cov_late.L1       0.00000      0.00000     0.000000      0.00000    0.00000
#> cov_late.L4       0.00000      0.00000     0.000000      0.00000    0.00000
#> cov_late.L3       0.00000      0.00000     0.000000      0.00000    0.00000
#>              cov_mid.L1 cov_mid.L4 cov_mid.L3 cov_late.L2 cov_late.L1
#> cov_early.L2    0.00000    0.00000    0.00000    0.000000      0.0000
#> cov_early.L1    0.00000    0.00000    0.00000    0.000000      0.0000
#> cov_early.L4    0.00000    0.00000    0.00000    0.000000      0.0000
#> cov_early.L3    0.00000    0.00000    0.00000    0.000000      0.0000
#> cov_mid.L2      0.00000    0.00000    0.00000    0.000000      0.0000
#> cov_mid.L1     10.16252    0.00000    0.00000    0.000000      0.0000
#> cov_mid.L4      0.00000   10.37202    0.00000    0.000000      0.0000
#> cov_mid.L3      0.00000    0.00000    7.41549    0.000000      0.0000
#> cov_late.L2     0.00000    0.00000    0.00000    6.517607      0.0000
#> cov_late.L1     0.00000    0.00000    0.00000    0.000000     13.1012
#> cov_late.L4     0.00000    0.00000    0.00000    0.000000      0.0000
#> cov_late.L3     0.00000    0.00000    0.00000    0.000000      0.0000
#>              cov_late.L4 cov_late.L3
#> cov_early.L2    0.000000     0.00000
#> cov_early.L1    0.000000     0.00000
#> cov_early.L4    0.000000     0.00000
#> cov_early.L3    0.000000     0.00000
#> cov_mid.L2      0.000000     0.00000
#> cov_mid.L1      0.000000     0.00000
#> cov_mid.L4      0.000000     0.00000
#> cov_mid.L3      0.000000     0.00000
#> cov_late.L2     0.000000     0.00000
#> cov_late.L1     0.000000     0.00000
#> cov_late.L4     3.569004     0.00000
#> cov_late.L3     0.000000    12.27148
#> 
#> $C_2
#>              cov_early.L2 cov_early.L1 cov_early.L4 cov_early.L3 cov_mid.L2
#> cov_early.L2     10.61294      0.00000     0.000000      0.00000   0.000000
#> cov_early.L1      0.00000     10.29417     0.000000      0.00000   0.000000
#> cov_early.L4      0.00000      0.00000     7.995691      0.00000   0.000000
#> cov_early.L3      0.00000      0.00000     0.000000      8.91606   0.000000
#> cov_mid.L2        0.00000      0.00000     0.000000      0.00000   8.955326
#> cov_mid.L1        0.00000      0.00000     0.000000      0.00000   0.000000
#> cov_mid.L4        0.00000      0.00000     0.000000      0.00000   0.000000
#> cov_mid.L3        0.00000      0.00000     0.000000      0.00000   0.000000
#> cov_late.L2       0.00000      0.00000     0.000000      0.00000   0.000000
#> cov_late.L1       0.00000      0.00000     0.000000      0.00000   0.000000
#> cov_late.L4       0.00000      0.00000     0.000000      0.00000   0.000000
#> cov_late.L3       0.00000      0.00000     0.000000      0.00000   0.000000
#>              cov_mid.L1 cov_mid.L4 cov_mid.L3 cov_late.L2 cov_late.L1
#> cov_early.L2   0.000000    0.00000   0.000000     0.00000     0.00000
#> cov_early.L1   0.000000    0.00000   0.000000     0.00000     0.00000
#> cov_early.L4   0.000000    0.00000   0.000000     0.00000     0.00000
#> cov_early.L3   0.000000    0.00000   0.000000     0.00000     0.00000
#> cov_mid.L2     0.000000    0.00000   0.000000     0.00000     0.00000
#> cov_mid.L1     9.754379    0.00000   0.000000     0.00000     0.00000
#> cov_mid.L4     0.000000   13.44659   0.000000     0.00000     0.00000
#> cov_mid.L3     0.000000    0.00000   3.407694     0.00000     0.00000
#> cov_late.L2    0.000000    0.00000   0.000000     7.14901     0.00000
#> cov_late.L1    0.000000    0.00000   0.000000     0.00000    13.15815
#> cov_late.L4    0.000000    0.00000   0.000000     0.00000     0.00000
#> cov_late.L3    0.000000    0.00000   0.000000     0.00000     0.00000
#>              cov_late.L4 cov_late.L3
#> cov_early.L2     0.00000     0.00000
#> cov_early.L1     0.00000     0.00000
#> cov_early.L4     0.00000     0.00000
#> cov_early.L3     0.00000     0.00000
#> cov_mid.L2       0.00000     0.00000
#> cov_mid.L1       0.00000     0.00000
#> cov_mid.L4       0.00000     0.00000
#> cov_mid.L3       0.00000     0.00000
#> cov_late.L2      0.00000     0.00000
#> cov_late.L1      0.00000     0.00000
#> cov_late.L4      7.85887     0.00000
#> cov_late.L3      0.00000    12.50978
#> 
#> $C_3
#>              cov_early.L2 cov_early.L1 cov_early.L4 cov_early.L3 cov_mid.L2
#> cov_early.L2     11.37098      0.00000     0.000000     0.000000    0.00000
#> cov_early.L1      0.00000     13.11305     0.000000     0.000000    0.00000
#> cov_early.L4      0.00000      0.00000     6.636361     0.000000    0.00000
#> cov_early.L3      0.00000      0.00000     0.000000     8.735088    0.00000
#> cov_mid.L2        0.00000      0.00000     0.000000     0.000000   10.47489
#> cov_mid.L1        0.00000      0.00000     0.000000     0.000000    0.00000
#> cov_mid.L4        0.00000      0.00000     0.000000     0.000000    0.00000
#> cov_mid.L3        0.00000      0.00000     0.000000     0.000000    0.00000
#> cov_late.L2       0.00000      0.00000     0.000000     0.000000    0.00000
#> cov_late.L1       0.00000      0.00000     0.000000     0.000000    0.00000
#> cov_late.L4       0.00000      0.00000     0.000000     0.000000    0.00000
#> cov_late.L3       0.00000      0.00000     0.000000     0.000000    0.00000
#>              cov_mid.L1 cov_mid.L4 cov_mid.L3 cov_late.L2 cov_late.L1
#> cov_early.L2   0.000000    0.00000   0.000000    0.000000     0.00000
#> cov_early.L1   0.000000    0.00000   0.000000    0.000000     0.00000
#> cov_early.L4   0.000000    0.00000   0.000000    0.000000     0.00000
#> cov_early.L3   0.000000    0.00000   0.000000    0.000000     0.00000
#> cov_mid.L2     0.000000    0.00000   0.000000    0.000000     0.00000
#> cov_mid.L1     9.952728    0.00000   0.000000    0.000000     0.00000
#> cov_mid.L4     0.000000   11.82215   0.000000    0.000000     0.00000
#> cov_mid.L3     0.000000    0.00000   8.479953    0.000000     0.00000
#> cov_late.L2    0.000000    0.00000   0.000000    8.089992     0.00000
#> cov_late.L1    0.000000    0.00000   0.000000    0.000000    13.16579
#> cov_late.L4    0.000000    0.00000   0.000000    0.000000     0.00000
#> cov_late.L3    0.000000    0.00000   0.000000    0.000000     0.00000
#>              cov_late.L4 cov_late.L3
#> cov_early.L2    0.000000     0.00000
#> cov_early.L1    0.000000     0.00000
#> cov_early.L4    0.000000     0.00000
#> cov_early.L3    0.000000     0.00000
#> cov_mid.L2      0.000000     0.00000
#> cov_mid.L1      0.000000     0.00000
#> cov_mid.L4      0.000000     0.00000
#> cov_mid.L3      0.000000     0.00000
#> cov_late.L2     0.000000     0.00000
#> cov_late.L1     0.000000     0.00000
#> cov_late.L4     7.036625     0.00000
#> cov_late.L3     0.000000    12.52641
#> 
#> $C_4
#>              cov_early.L2 cov_early.L1 cov_early.L4 cov_early.L3 cov_mid.L2
#> cov_early.L2      11.2961      0.00000     0.000000     0.000000    0.00000
#> cov_early.L1       0.0000     10.80589     0.000000     0.000000    0.00000
#> cov_early.L4       0.0000      0.00000     9.098493     0.000000    0.00000
#> cov_early.L3       0.0000      0.00000     0.000000     4.559901    0.00000
#> cov_mid.L2         0.0000      0.00000     0.000000     0.000000   10.56905
#> cov_mid.L1         0.0000      0.00000     0.000000     0.000000    0.00000
#> cov_mid.L4         0.0000      0.00000     0.000000     0.000000    0.00000
#> cov_mid.L3         0.0000      0.00000     0.000000     0.000000    0.00000
#> cov_late.L2        0.0000      0.00000     0.000000     0.000000    0.00000
#> cov_late.L1        0.0000      0.00000     0.000000     0.000000    0.00000
#> cov_late.L4        0.0000      0.00000     0.000000     0.000000    0.00000
#> cov_late.L3        0.0000      0.00000     0.000000     0.000000    0.00000
#>              cov_mid.L1 cov_mid.L4 cov_mid.L3 cov_late.L2 cov_late.L1
#> cov_early.L2    0.00000    0.00000   0.000000    0.000000     0.00000
#> cov_early.L1    0.00000    0.00000   0.000000    0.000000     0.00000
#> cov_early.L4    0.00000    0.00000   0.000000    0.000000     0.00000
#> cov_early.L3    0.00000    0.00000   0.000000    0.000000     0.00000
#> cov_mid.L2      0.00000    0.00000   0.000000    0.000000     0.00000
#> cov_mid.L1     13.45513    0.00000   0.000000    0.000000     0.00000
#> cov_mid.L4      0.00000   11.40982   0.000000    0.000000     0.00000
#> cov_mid.L3      0.00000    0.00000   8.221683    0.000000     0.00000
#> cov_late.L2     0.00000    0.00000   0.000000    8.556541     0.00000
#> cov_late.L1     0.00000    0.00000   0.000000    0.000000    14.92923
#> cov_late.L4     0.00000    0.00000   0.000000    0.000000     0.00000
#> cov_late.L3     0.00000    0.00000   0.000000    0.000000     0.00000
#>              cov_late.L4 cov_late.L3
#> cov_early.L2    0.000000     0.00000
#> cov_early.L1    0.000000     0.00000
#> cov_early.L4    0.000000     0.00000
#> cov_early.L3    0.000000     0.00000
#> cov_mid.L2      0.000000     0.00000
#> cov_mid.L1      0.000000     0.00000
#> cov_mid.L4      0.000000     0.00000
#> cov_mid.L3      0.000000     0.00000
#> cov_late.L2     0.000000     0.00000
#> cov_late.L1     0.000000     0.00000
#> cov_late.L4     8.362115     0.00000
#> cov_late.L3     0.000000    12.52706
#> 
#> $C_5
#>              cov_early.L2 cov_early.L1 cov_early.L4 cov_early.L3 cov_mid.L2
#> cov_early.L2      26.8613      0.00000      0.00000       0.0000    0.00000
#> cov_early.L1       0.0000     25.46037      0.00000       0.0000    0.00000
#> cov_early.L4       0.0000      0.00000     15.80356       0.0000    0.00000
#> cov_early.L3       0.0000      0.00000      0.00000      15.4666    0.00000
#> cov_mid.L2         0.0000      0.00000      0.00000       0.0000   22.92828
#> cov_mid.L1         0.0000      0.00000      0.00000       0.0000    0.00000
#> cov_mid.L4         0.0000      0.00000      0.00000       0.0000    0.00000
#> cov_mid.L3         0.0000      0.00000      0.00000       0.0000    0.00000
#> cov_late.L2        0.0000      0.00000      0.00000       0.0000    0.00000
#> cov_late.L1        0.0000      0.00000      0.00000       0.0000    0.00000
#> cov_late.L4        0.0000      0.00000      0.00000       0.0000    0.00000
#> cov_late.L3        0.0000      0.00000      0.00000       0.0000    0.00000
#>              cov_mid.L1 cov_mid.L4 cov_mid.L3 cov_late.L2 cov_late.L1
#> cov_early.L2    0.00000    0.00000    0.00000     0.00000     0.00000
#> cov_early.L1    0.00000    0.00000    0.00000     0.00000     0.00000
#> cov_early.L4    0.00000    0.00000    0.00000     0.00000     0.00000
#> cov_early.L3    0.00000    0.00000    0.00000     0.00000     0.00000
#> cov_mid.L2      0.00000    0.00000    0.00000     0.00000     0.00000
#> cov_mid.L1     22.44793    0.00000    0.00000     0.00000     0.00000
#> cov_mid.L4      0.00000   27.48263    0.00000     0.00000     0.00000
#> cov_mid.L3      0.00000    0.00000   14.65999     0.00000     0.00000
#> cov_late.L2     0.00000    0.00000    0.00000    16.00021     0.00000
#> cov_late.L1     0.00000    0.00000    0.00000     0.00000    49.64581
#> cov_late.L4     0.00000    0.00000    0.00000     0.00000     0.00000
#> cov_late.L3     0.00000    0.00000    0.00000     0.00000     0.00000
#>              cov_late.L4 cov_late.L3
#> cov_early.L2     0.00000     0.00000
#> cov_early.L1     0.00000     0.00000
#> cov_early.L4     0.00000     0.00000
#> cov_early.L3     0.00000     0.00000
#> cov_mid.L2       0.00000     0.00000
#> cov_mid.L1       0.00000     0.00000
#> cov_mid.L4       0.00000     0.00000
#> cov_mid.L3       0.00000     0.00000
#> cov_late.L2      0.00000     0.00000
#> cov_late.L1      0.00000     0.00000
#> cov_late.L4     15.29367     0.00000
#> cov_late.L3      0.00000    41.01066
#> 
#> $C_6
#>              cov_early.L2 cov_early.L1 cov_early.L4 cov_early.L3 cov_mid.L2
#> cov_early.L2     28.68768      0.00000      0.00000      0.00000     0.0000
#> cov_early.L1      0.00000     24.80685      0.00000      0.00000     0.0000
#> cov_early.L4      0.00000      0.00000     15.34661      0.00000     0.0000
#> cov_early.L3      0.00000      0.00000      0.00000     15.06992     0.0000
#> cov_mid.L2        0.00000      0.00000      0.00000      0.00000    22.1826
#> cov_mid.L1        0.00000      0.00000      0.00000      0.00000     0.0000
#> cov_mid.L4        0.00000      0.00000      0.00000      0.00000     0.0000
#> cov_mid.L3        0.00000      0.00000      0.00000      0.00000     0.0000
#> cov_late.L2       0.00000      0.00000      0.00000      0.00000     0.0000
#> cov_late.L1       0.00000      0.00000      0.00000      0.00000     0.0000
#> cov_late.L4       0.00000      0.00000      0.00000      0.00000     0.0000
#> cov_late.L3       0.00000      0.00000      0.00000      0.00000     0.0000
#>              cov_mid.L1 cov_mid.L4 cov_mid.L3 cov_late.L2 cov_late.L1
#> cov_early.L2    0.00000     0.0000    0.00000     0.00000     0.00000
#> cov_early.L1    0.00000     0.0000    0.00000     0.00000     0.00000
#> cov_early.L4    0.00000     0.0000    0.00000     0.00000     0.00000
#> cov_early.L3    0.00000     0.0000    0.00000     0.00000     0.00000
#> cov_mid.L2      0.00000     0.0000    0.00000     0.00000     0.00000
#> cov_mid.L1     22.76154     0.0000    0.00000     0.00000     0.00000
#> cov_mid.L4      0.00000    29.4899    0.00000     0.00000     0.00000
#> cov_mid.L3      0.00000     0.0000   13.85003     0.00000     0.00000
#> cov_late.L2     0.00000     0.0000    0.00000    12.97552     0.00000
#> cov_late.L1     0.00000     0.0000    0.00000     0.00000    69.78387
#> cov_late.L4     0.00000     0.0000    0.00000     0.00000     0.00000
#> cov_late.L3     0.00000     0.0000    0.00000     0.00000     0.00000
#>              cov_late.L4 cov_late.L3
#> cov_early.L2     0.00000     0.00000
#> cov_early.L1     0.00000     0.00000
#> cov_early.L4     0.00000     0.00000
#> cov_early.L3     0.00000     0.00000
#> cov_mid.L2       0.00000     0.00000
#> cov_mid.L1       0.00000     0.00000
#> cov_mid.L4       0.00000     0.00000
#> cov_mid.L3       0.00000     0.00000
#> cov_late.L2      0.00000     0.00000
#> cov_late.L1      0.00000     0.00000
#> cov_late.L4     13.74047     0.00000
#> cov_late.L3      0.00000    34.58792
#> 
#> $C_7
#>              cov_early.L2 cov_early.L1 cov_early.L4 cov_early.L3 cov_mid.L2
#> cov_early.L2     11.34156      0.00000     0.000000     0.000000    0.00000
#> cov_early.L1      0.00000     13.55361     0.000000     0.000000    0.00000
#> cov_early.L4      0.00000      0.00000     6.305458     0.000000    0.00000
#> cov_early.L3      0.00000      0.00000     0.000000     8.481697    0.00000
#> cov_mid.L2        0.00000      0.00000     0.000000     0.000000   10.36295
#> cov_mid.L1        0.00000      0.00000     0.000000     0.000000    0.00000
#> cov_mid.L4        0.00000      0.00000     0.000000     0.000000    0.00000
#> cov_mid.L3        0.00000      0.00000     0.000000     0.000000    0.00000
#> cov_late.L2       0.00000      0.00000     0.000000     0.000000    0.00000
#> cov_late.L1       0.00000      0.00000     0.000000     0.000000    0.00000
#> cov_late.L4       0.00000      0.00000     0.000000     0.000000    0.00000
#> cov_late.L3       0.00000      0.00000     0.000000     0.000000    0.00000
#>              cov_mid.L1 cov_mid.L4 cov_mid.L3 cov_late.L2 cov_late.L1
#> cov_early.L2    0.00000    0.00000    0.00000    0.000000     0.00000
#> cov_early.L1    0.00000    0.00000    0.00000    0.000000     0.00000
#> cov_early.L4    0.00000    0.00000    0.00000    0.000000     0.00000
#> cov_early.L3    0.00000    0.00000    0.00000    0.000000     0.00000
#> cov_mid.L2      0.00000    0.00000    0.00000    0.000000     0.00000
#> cov_mid.L1     10.15705    0.00000    0.00000    0.000000     0.00000
#> cov_mid.L4      0.00000   11.40124    0.00000    0.000000     0.00000
#> cov_mid.L3      0.00000    0.00000    9.37083    0.000000     0.00000
#> cov_late.L2     0.00000    0.00000    0.00000    7.243544     0.00000
#> cov_late.L1     0.00000    0.00000    0.00000    0.000000    18.93351
#> cov_late.L4     0.00000    0.00000    0.00000    0.000000     0.00000
#> cov_late.L3     0.00000    0.00000    0.00000    0.000000     0.00000
#>              cov_late.L4 cov_late.L3
#> cov_early.L2    0.000000     0.00000
#> cov_early.L1    0.000000     0.00000
#> cov_early.L4    0.000000     0.00000
#> cov_early.L3    0.000000     0.00000
#> cov_mid.L2      0.000000     0.00000
#> cov_mid.L1      0.000000     0.00000
#> cov_mid.L4      0.000000     0.00000
#> cov_mid.L3      0.000000     0.00000
#> cov_late.L2     0.000000     0.00000
#> cov_late.L1     0.000000     0.00000
#> cov_late.L4     8.374095     0.00000
#> cov_late.L3     0.000000    12.52634
#> 
#> $C_8
#>              cov_early.L2 cov_early.L1 cov_early.L4 cov_early.L3 cov_mid.L2
#> cov_early.L2     13.72974      0.00000     0.000000     0.000000    0.00000
#> cov_early.L1      0.00000     10.93231     0.000000     0.000000    0.00000
#> cov_early.L4      0.00000      0.00000     9.627852     0.000000    0.00000
#> cov_early.L3      0.00000      0.00000     0.000000     9.419364    0.00000
#> cov_mid.L2        0.00000      0.00000     0.000000     0.000000   11.38538
#> cov_mid.L1        0.00000      0.00000     0.000000     0.000000    0.00000
#> cov_mid.L4        0.00000      0.00000     0.000000     0.000000    0.00000
#> cov_mid.L3        0.00000      0.00000     0.000000     0.000000    0.00000
#> cov_late.L2       0.00000      0.00000     0.000000     0.000000    0.00000
#> cov_late.L1       0.00000      0.00000     0.000000     0.000000    0.00000
#> cov_late.L4       0.00000      0.00000     0.000000     0.000000    0.00000
#> cov_late.L3       0.00000      0.00000     0.000000     0.000000    0.00000
#>              cov_mid.L1 cov_mid.L4 cov_mid.L3 cov_late.L2 cov_late.L1
#> cov_early.L2    0.00000    0.00000   0.000000    0.000000     0.00000
#> cov_early.L1    0.00000    0.00000   0.000000    0.000000     0.00000
#> cov_early.L4    0.00000    0.00000   0.000000    0.000000     0.00000
#> cov_early.L3    0.00000    0.00000   0.000000    0.000000     0.00000
#> cov_mid.L2      0.00000    0.00000   0.000000    0.000000     0.00000
#> cov_mid.L1     10.68254    0.00000   0.000000    0.000000     0.00000
#> cov_mid.L4      0.00000   11.39224   0.000000    0.000000     0.00000
#> cov_mid.L3      0.00000    0.00000   8.826938    0.000000     0.00000
#> cov_late.L2     0.00000    0.00000   0.000000    3.796296     0.00000
#> cov_late.L1     0.00000    0.00000   0.000000    0.000000    13.16397
#> cov_late.L4     0.00000    0.00000   0.000000    0.000000     0.00000
#> cov_late.L3     0.00000    0.00000   0.000000    0.000000     0.00000
#>              cov_late.L4 cov_late.L3
#> cov_early.L2     0.00000     0.00000
#> cov_early.L1     0.00000     0.00000
#> cov_early.L4     0.00000     0.00000
#> cov_early.L3     0.00000     0.00000
#> cov_mid.L2       0.00000     0.00000
#> cov_mid.L1       0.00000     0.00000
#> cov_mid.L4       0.00000     0.00000
#> cov_mid.L3       0.00000     0.00000
#> cov_late.L2      0.00000     0.00000
#> cov_late.L1      0.00000     0.00000
#> cov_late.L4     11.19266     0.00000
#> cov_late.L3      0.00000    12.52703
#> 
#> $C_9
#>              cov_early.L2 cov_early.L1 cov_early.L4 cov_early.L3 cov_mid.L2
#> cov_early.L2     11.20425       0.0000     0.000000     0.000000    0.00000
#> cov_early.L1      0.00000      10.9508     0.000000     0.000000    0.00000
#> cov_early.L4      0.00000       0.0000     8.546684     0.000000    0.00000
#> cov_early.L3      0.00000       0.0000     0.000000     9.088426    0.00000
#> cov_mid.L2        0.00000       0.0000     0.000000     0.000000   10.64808
#> cov_mid.L1        0.00000       0.0000     0.000000     0.000000    0.00000
#> cov_mid.L4        0.00000       0.0000     0.000000     0.000000    0.00000
#> cov_mid.L3        0.00000       0.0000     0.000000     0.000000    0.00000
#> cov_late.L2       0.00000       0.0000     0.000000     0.000000    0.00000
#> cov_late.L1       0.00000       0.0000     0.000000     0.000000    0.00000
#> cov_late.L4       0.00000       0.0000     0.000000     0.000000    0.00000
#> cov_late.L3       0.00000       0.0000     0.000000     0.000000    0.00000
#>              cov_mid.L1 cov_mid.L4 cov_mid.L3 cov_late.L2 cov_late.L1
#> cov_early.L2    0.00000    0.00000    0.00000     0.00000     0.00000
#> cov_early.L1    0.00000    0.00000    0.00000     0.00000     0.00000
#> cov_early.L4    0.00000    0.00000    0.00000     0.00000     0.00000
#> cov_early.L3    0.00000    0.00000    0.00000     0.00000     0.00000
#> cov_mid.L2      0.00000    0.00000    0.00000     0.00000     0.00000
#> cov_mid.L1     11.29461    0.00000    0.00000     0.00000     0.00000
#> cov_mid.L4      0.00000   11.36996    0.00000     0.00000     0.00000
#> cov_mid.L3      0.00000    0.00000    9.06309     0.00000     0.00000
#> cov_late.L2     0.00000    0.00000    0.00000     9.63419     0.00000
#> cov_late.L1     0.00000    0.00000    0.00000     0.00000    13.16666
#> cov_late.L4     0.00000    0.00000    0.00000     0.00000     0.00000
#> cov_late.L3     0.00000    0.00000    0.00000     0.00000     0.00000
#>              cov_late.L4 cov_late.L3
#> cov_early.L2    0.000000     0.00000
#> cov_early.L1    0.000000     0.00000
#> cov_early.L4    0.000000     0.00000
#> cov_early.L3    0.000000     0.00000
#> cov_mid.L2      0.000000     0.00000
#> cov_mid.L1      0.000000     0.00000
#> cov_mid.L4      0.000000     0.00000
#> cov_mid.L3      0.000000     0.00000
#> cov_late.L2     0.000000     0.00000
#> cov_late.L1     0.000000     0.00000
#> cov_late.L4     8.296323     0.00000
#> cov_late.L3     0.000000    12.84321
#> 
#> $C_10
#>              cov_early.L2 cov_early.L1 cov_early.L4 cov_early.L3 cov_mid.L2
#> cov_early.L2     11.38747      0.00000     0.000000     0.000000    0.00000
#> cov_early.L1      0.00000     15.81282     0.000000     0.000000    0.00000
#> cov_early.L4      0.00000      0.00000     9.659709     0.000000    0.00000
#> cov_early.L3      0.00000      0.00000     0.000000     7.693014    0.00000
#> cov_mid.L2        0.00000      0.00000     0.000000     0.000000   10.66775
#> cov_mid.L1        0.00000      0.00000     0.000000     0.000000    0.00000
#> cov_mid.L4        0.00000      0.00000     0.000000     0.000000    0.00000
#> cov_mid.L3        0.00000      0.00000     0.000000     0.000000    0.00000
#> cov_late.L2       0.00000      0.00000     0.000000     0.000000    0.00000
#> cov_late.L1       0.00000      0.00000     0.000000     0.000000    0.00000
#> cov_late.L4       0.00000      0.00000     0.000000     0.000000    0.00000
#> cov_late.L3       0.00000      0.00000     0.000000     0.000000    0.00000
#>              cov_mid.L1 cov_mid.L4 cov_mid.L3 cov_late.L2 cov_late.L1
#> cov_early.L2    0.00000    0.00000   0.000000    0.000000     0.00000
#> cov_early.L1    0.00000    0.00000   0.000000    0.000000     0.00000
#> cov_early.L4    0.00000    0.00000   0.000000    0.000000     0.00000
#> cov_early.L3    0.00000    0.00000   0.000000    0.000000     0.00000
#> cov_mid.L2      0.00000    0.00000   0.000000    0.000000     0.00000
#> cov_mid.L1     10.63265    0.00000   0.000000    0.000000     0.00000
#> cov_mid.L4      0.00000   11.39569   0.000000    0.000000     0.00000
#> cov_mid.L3      0.00000    0.00000   9.378185    0.000000     0.00000
#> cov_late.L2     0.00000    0.00000   0.000000    9.536634     0.00000
#> cov_late.L1     0.00000    0.00000   0.000000    0.000000    13.16637
#> cov_late.L4     0.00000    0.00000   0.000000    0.000000     0.00000
#> cov_late.L3     0.00000    0.00000   0.000000    0.000000     0.00000
#>              cov_late.L4 cov_late.L3
#> cov_early.L2    0.000000     0.00000
#> cov_early.L1    0.000000     0.00000
#> cov_early.L4    0.000000     0.00000
#> cov_early.L3    0.000000     0.00000
#> cov_mid.L2      0.000000     0.00000
#> cov_mid.L1      0.000000     0.00000
#> cov_mid.L4      0.000000     0.00000
#> cov_mid.L3      0.000000     0.00000
#> cov_late.L2     0.000000     0.00000
#> cov_late.L1     0.000000     0.00000
#> cov_late.L4     6.991775     0.00000
#> cov_late.L3     0.000000    15.05481
#>